factor our MemCheckinStrat; rename to FingerprintStrategy
