+++ src/lasso/mod.rs
@@ -1,3 +1,4 @@
 pub mod densified;
 pub mod memory_checking;
 pub mod surge;
+pub mod fingerprint_strategy;
